Added Vehicle model. Code cleaning and dividing TrajectoryFollowers to separate files.
